some reworkings

+++ examples/minwebgl/hexagonal_map/shaders/main.frag
@@ -12,7 +12,7 @@ const vec3[] PALETTE = vec3[]
 );
 
 in vec2 v_tex_coord;
-flat in float v_player_id;
+flat in int v_player_id;
 
 uniform sampler2D u_sprite_sheet;
 
@@ -21,7 +21,7 @@ layout ( location = 0 ) out vec4 frag_color;
 void main()
 {
   vec4 color = texture( u_sprite_sheet, v_tex_coord );
-  vec3 player_color = PALETTE[ int( v_player_id ) ];
+  vec3 player_color = PALETTE[ v_player_id ];
 
   frag_color = vec4( mix( color.rgb , player_color, 1.0 ), 1.0 );
 }
